% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_quant_report.R
\name{export_quant_report}
\alias{export_quant_report}
\title{Analyze quant fluorescence data}
\usage{
export_quant_report(quant_dna)
}
\arguments{
\item{quant_dna}{list. Output of \code{quant_dna}}
}
\value{
Generates an excel file of the quant report.
}
\description{
Wrangles and exports the quant data in a format used by the Genomics Core Laboratory (GCL).

This is step 4 of 4 in a series of functions for analyzing fluorescent data.
}
\examples{
Import data files
raw_data <- system.file("extdata", "raw_data.csv", package = "tamuccGCL")
plate_map <- system.file("extdata", "plate_map.csv", package = "tamuccGCL")

quant_data <- load_quant_files(raw_data, plate_map)
trained_model <- train_quant_standards (quant_data)
quant_report <- quant_dna(quant_data = quant_data, trained_model = trained_model)
export_quant_report (quant_report)

}
