% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quant_files.R
\name{load_quant_files}
\alias{load_quant_files}
\title{Analyze flourescent data}
\usage{
load_quant_files(raw_data, plate_map)
}
\arguments{
\item{raw_data}{Raw plate reader data saved as a comma-separated value (csv) file. This is the file exported from the SpectraMax software. It contains the well and fluorescence information. This file should not be manipulated once it has been exported.}

\item{plate_map}{This contains the information on how the samples were laid out in the 364-well plate and its associated metadata. This file follows a specific data structure. The following columns are expected:
\itemize{
\item plate_id: Name of the sample plate/strip that has been assayed. All standards used in the assay should be named "standard".
\item sample_id: Name of the samples in the plate. This usually corresponds to well position. Standards used in this assay should be named based on the concentration (ng/ul) used.
\item replicate: A numeric value containing the replicate used in the assay. A value of 2 is expected for "double-quants"
\item quant_row: A letter indicating the row of the sample in the quant plate.
\item quant_column: A number indicating the column of the sample in the quant plate.
\item sample_volume: A number indicating the volume (uL) of samples and standards used in the assay.
}}
}
\value{
Returns a list containing two data frames: sample and standard
}
\description{
Prepare the raw data files for fluorescence analysis. This function reads the raw data and plate map, validates the file format and data structure, merges the files, and then prepares a list that separates the samples from the standards.

This is the first step in a series of functions for analyzing fluorescent data.
}
\details{
Additional details...
}
\examples{
# Import data files
raw_data <- system.file("extdata", "raw_data.csv", package = "tamuccGCL")
plate_map <- system.file("extdata", "plate_map.csv", package = "tamuccGCL")

quant_data <- load_quant_files(raw_data, plate_map)
}
