\name{gclTools}
\alias{gclTools}
\docType{package}
\title{gclTools: Automating TAMU-CC GCL workflow}
\description{
  The gclTools package provides a suite of functions designed for analyzing and visualizing GCL data. This addresses computational workflows in TAMU-CC Genomics Core Laboratory.
}
\author{
  Kevin Labrador [aut, cre] \cr
  Christopher Bird [ctb, rev] \cr
}
\seealso{
  \code{\link[=load_quant_files]{load_quant_files()}},
  \code{\link[=train_quant_standards]{train_quant_standards()}},
  \code{\link[=quant_dna]{quant_dna()}},
  \code{\link[=export_quant_report]{export_quant_report()}}
}
\keyword{package}

